refactor(bibliography): 消除 0.4.x 遗留依赖环，拆分题录域，包内零循环依赖
## 环的根因与消除

test_no_new_import_cycles_appear 冻结的
database ↔ text_alignment ↔ calibration_library ↔ bibliographic_metadata
是包内最后一条环。根因是 bibliographic_metadata 为了一个纯函数
(paragraph_payload_for_storage) 顶层依赖整个 database 模块。

该函数与 paragraph_from_database_row 只做「SQLite 行 ↔ 规范字典」的形状转换，
不含任何领域概念，属 persistence 而非 database：连同三个段落列常量下沉到
persistence/paragraph_payload.py，database 原样重新导出。bibliographic_metadata
改从 persistence 取，这条边不复存在。

known 已由那一条环改为空集：包内现在零循环依赖（含函数内懒 import）。

## 题录域拆分

bibliographic_metadata 2384 → 1530 行，按「共用底座 + 按文献类型分」四拆：

  bibliographic_values        315 行  取值校验、人名/出版者清洗、候选打分（最底层）
  bibliographic_thesis        287 行  学位论文封面抽取
  bibliographic_journal       342 行  期刊论文判别与抽取
  bibliographic_marx_engels   165 行  马恩全集卷次—年份对照表

分区经 AST 分层方向检查验证，零违规；同层三个抽取器之间无互相依赖。公开面
（METADATA_FIELDS、canonical_metadata、normalize_doi、marx_engels_* 等）
仍从 bibliographic_metadata 重新导出，外部调用面零变化。

## 顺带解锁

上一轮因会拉长该环而退回的对齐配方快照，现已可安全移出：
alignment_snapshots.py 142 行，text_alignment 2484 → 2373。

## 门禁（只降不升）

  bibliographic_metadata  2400 → 1550     text_alignment  2500 → 2400
  database                1500 → 1325     新增 5 个模块上限

全量 unittest discover 1981 项通过（1 跳过）。
